# openapi.yaml
...
openapi: 3.0.0
info:
  title: Telecloud Server
  version: "1.0"
  description: Telecloud Server provides an API to turn a Telegram chanel into a
    flexible file system using folders and tags.
  termsOfService: ""
  contact:
    name: estebanthi
    email: contact@esteban-thilliez.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-logo:
    url: ""
servers:
  - url: https://api.example.io
paths:
  /files:
    summary: Interactions with the file system
    get:
      summary: Download files
      description: |-
        Downloads the files according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "123456789"
                  - "987654321"
          description: OK
    post:
      summary: Upload files
      description: Uploads the files passed in parameters
      tags:
        - Files
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Files to upload
                  example:
                    - file1
                    - file2
                data:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        description: Type of the file
                        example: jpeg
                      size:
                        type: integer
                        description: Size of the file
                        example: 123456
                      created_at:
                        type: string
                        format: date-time
                        description: Creation date of the file
                        example: 2020-01-01T00:00:00Z
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "123456789"
                  - "987654321"
                description: IDs of the uploaded files
          description: OK
    delete:
      summary: Delete files
      description: |-
        Delete the files according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
                description: Deleted files
          description: OK
  /files/id:
    summary: Interactions with files' IDs
    get:
      summary: Get files IDs
      description: |-
        Get the files IDs according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "123456789"
                  - "987654321"
          description: OK
  /files/meta:
    summary: Interactions with files' representation
    get:
      summary: Get files meta
      description: |-
        Get the files meta according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/File"
          description: OK
    patch:
      summary: Update files meta
      description: >-
        Update the files' representation according to the filter passed in
        parameters.


        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  example:
                    - pictures
                    - cars
                  items:
                    type: string
                directory:
                  type: string
                  example: "123456789"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                    - "123456789"
                    - "987654321"
                description: Updated files
          description: OK
  /files/meta/tags:
    summary: Interactions with files' tags
    post:
      summary: Add tags
      description: |-
        Add tags to files according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  example:
                    - "2023"
                  items:
                    type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                    type: string
                example:
                    - "123456789"
                    - "987654321"
                description: Updated files
          description: OK
    patch:
      summary: Delete some tags
      description: |-
        Remove given tags in files according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  example:
                    - cars
                  items:
                    type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                    type: string
                example:
                    - "123456789"
                    - "987654321"
                description: Updated files
          description: OK
    delete:
      summary: Delete all tags
      description: |-
        Delete files' tags according to the filter passed in parameters.

        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                    type: string
                example:
                    - "123456789"
                    - "987654321"
                description: Updated files
          description: OK
  /files/meta/directory:
    summary: Interactions with files' directories
    delete:
      summary: Delete files' directory
      description: >-
        Delete files' directory according to the filter passed in parameters.
        
        **Note**: it is an `AND` filter.
      tags:
        - Files
      parameters:
        - name: tags
          description: Tags associated with the files
          example:
            - pictures
            - cars
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: types
          description: Types of the files
          example:
            - jpeg
            - png
          schema:
            type: array
            items:
              type: string
          in: query
          required: false
          allowEmptyValue: true
        - name: directories
          description: Directories ID of the files
          in: query
          required: false
          example:
            - "123456789"
            - "987654321"
          schema:
            type: array
            items:
              type: string
          allowEmptyValue: true
      responses:
            "200":
                content:
                  application/json:
                    schema:
                      type: array
                      description: IDs of the modified files
                      items:
                        type: string
                      example:
                        - "123456789"
                        - "987654321"
                description: OK
  "/files/{file_id}":
    summary: Interactions with the file system for individual files
    parameters:
      - name: file_id
        in: path
        schema:
          type: string
          example: "123456789"
        required: true
    get:
      summary: Download file
      description: Downloads the given file.
      tags:
        - Files
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
                example: "123456789"
                description: ID of the downloaded file
          description: OK
    delete:
      summary: Delete file
      description: Delete the given file
      tags:
        - Files
      responses:
        "200":
          content:
            application/json:
              schema:
                type: string
                example: "123456789"
                description: ID of the deleted file
          description: OK
...
  "/files/{file_id}/meta/directory":
    summary: Interactions with file directory
    parameters:
      - name: file_id
        in: path
        schema:
          type: string
          example: "123456789"
        required: true
    delete:
        summary: Delete file directory
        description: Delete the directory of the given file.
        tags:
            - Files
        responses:
            "200":
              content:
                application/json:
                  schema:
                    $ref: "#/components/schemas/File"
              description: OK
...
components:
  schemas:
    File:
      type: object
      required:
        - _id
        - name
        - size
        - extension
      properties:
        _id:
          type: string
          example: "123456789"
        name:
          type: string
          example: My file
        size:
          type: integer
          example: 34567
        extension:
          type: string
          example: pdf
        directory:
          type: string
          example: "123456789"
        tags:
          type: array
          items:
            type: string
          example:
            - pictures
            - cars
        created_at:
          type: string
          example: ""
        uploaded_at:
          type: string
          example: ""
        chunks:
          type: array
          items:
            type: string
          example:
            - "123456789"
            - "987654321"
          description: Chunks composing the file
...